refactor: extract NavigationHelpers.ts from IOU/index.ts
Extracts post-expense-creation navigation helpers (dismissModalAndOpenReportInInboxTab,
highlightTransactionOnSearchRouteIfNeeded, handleNavigateAfterExpenseCreate) into a
dedicated module to reduce the size of IOU/index.ts.

Function bodies are byte-identical to their pre-move state on main, with the single
required translation of direct allTransactions reads to getAllTransactions() calls
(needed because allTransactions is module-private to index.ts).

Part of Issue #72804 (break up src/libs/actions/IOU/).
